requirements.txt
discord.py-self==2.0.0
aiohttp>=3.8.0
asyncio
